Don't reset state when seeing definition
We might see the use of an object before its definition. This is the
case when dealing with foo_test packages that make use of objects
declared in foo's (not foo_test's) tests.

No test for now because the testing infrastructure can't handle
multi-file packages, and even less a mixture of packages.

Closes gh-6