Ignore identifiers generated by cgo
There's no point in reporting identifiers generated by cgo, as the user
cannot act on them.

The documentation in
https://github.com/golang/tools/blob/master/go/loader/cgo.go suggests
that C types will be declared in a file called "C". However, we've also
encountered tokens with no filename attached at all. We ignore both
tokens in the C file and those with no file name.

Closes gh-31