Remove dubious code from processAST
At first sight, the code appears to be a noop. One switch case is doing
nothing, another one is buggy and marks a node as used by itself.

Theoretically, removing this code should not have any effect.
Realistically, removing it does change behaviour in a corner case.
Consider the following example:

	typ := reflect.TypeOf(struct {
		d []uint32 `reflect:"TAG"`
	}{}).Field(0).Type

With the code we're removing, the anonymous struct will be added to the
graph. Later, we iterate over all structs in the graph, and mark their
fields as quiet if the struct itself is unused. Without the code, we'll
never mark the field as quiet, and it will be reported as unused.

Both behaviours are corner cases in their own right and will probably
lead to odd behaviour when encountering odd code. We'll err on the side
of having less dubious code and fixing concrete issues once people with
realistic code report them.